removed pt::array usages (lots of missing features) WorldComponent   fixed uninitalized pointer Entity   removed CreateContext() and DestroyContext()   added Spawn() and Despawn() DrawableComponent   added CreateContext and DestroyContext

+++ src/sdlcontrol.cpp
@@ -95,14 +95,14 @@ SetMainRenderer(SDL_Renderer *renderer)
 }
 
 
-float SDLControl::
+uint32_t SDLControl::
 GetMainWindowWidth() const
 {
     return mMainWindowWidth;
 }
 
 
-float SDLControl::
+uint32_t SDLControl::
 GetMainWindowHeight() const
 {
     return mMainWindowHeight;
